add board Err Response

+++ src/controller/boardController.js
@@ -12,6 +12,7 @@ exports.createPost = async function (req, res) {
     if (!postName) return res.send(response(baseResponse.POST_NAME_EMPTY));
     if (!title) return res.send(response(baseResponse.POST_TITLE_EMPTY));
     if (!mainText) return res.send(response(baseResponse.POST_MAIN_TEXT_EMPTY));
+    if (mainText.length > 65535) return res.send(response(baseResponse.POST_MAIN_TEXT_TOO_LONG));
 
     const createPostResponse = await boardService.createPost(postName, userId, title, mainText, postType);
 
@@ -29,9 +30,21 @@ exports.getPosts = async function (req, res) {
 exports.updatePost = async function (req, res) {
     const {postIdx} = req.params;
     const {title, mainText} = req.body;
+    const userIdFromJWT = req.verifiedToken.userId; // JWT에서 userId 가져오기
 
     if (!title) return res.send(response(baseResponse.POST_TITLE_EMPTY));
     if (!mainText) return res.send(response(baseResponse.POST_MAIN_TEXT_EMPTY));
+
+    // 게시글 내용이 너무 긴 경우
+    if (mainText.length > 65535) return res.send(response(baseResponse.POST_MAIN_TEXT_TOO_LONG));
+
+    const post = await boardProvider.retrievePost(postIdx); // 게시글 조회
+
+    // 게시글이 존재하지 않는 경우
+    if (!post) return res.send(response(baseResponse.POST_NOT_FOUND));
+
+    // JWT의 userId와 게시글의 작성자가 일치하지 않는 경우
+    if (post.userId !== userIdFromJWT) return res.send(response(baseResponse.POST_NOT_WRITER));
 
     const updatePostResponse = await boardService.updatePost(postIdx, title, mainText);
 
@@ -41,6 +54,15 @@ exports.updatePost = async function (req, res) {
 // 게시글 삭제
 exports.deletePost = async function (req, res) {
     const {postIdx} = req.params;
+    const userIdFromJWT = req.verifiedToken.userId; // JWT에서 userId 가져오기
+
+    const post = await boardProvider.retrievePost(postIdx); // 게시글 조회
+
+    // 게시글이 존재하지 않는 경우
+    if (!post) return res.send(response(baseResponse.POST_NOT_FOUND));
+
+    // JWT의 userId와 게시글의 작성자가 일치하지 않는 경우
+    if (post.userId !== userIdFromJWT) return res.send(response(baseResponse.POST_NOT_WRITER));
 
     const deletePostResponse = await boardService.deletePost(postIdx);
